requirements.txt
alembic==1.10.2
beautifulsoup4==4.11.2
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.3
Flask-HTTPAuth==4.7.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
PyJWT==2.6.0
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4
SQLAlchemy==2.0.6
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
